+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:42841/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:45769/jest?","mongoDBName":"jest"}
