+++ src/presentation/controllers/login/login.spec.ts
@@ -1,7 +1,7 @@
-import { Authentication } from '../../../domain/useCases/authentication'
-import { InvalidParamError, MissingParamError } from '../../errors'
+import { EmailValidator, HttpRequest, Authentication } from './login-protocols'
 import { badRequest, serverError, unauthorized } from '../../helpers/http-helper'
-import { EmailValidator, HttpRequest } from '../../protocols'
+import { InvalidParamError, MissingParamError } from '../../errors'
+
 import { LoginController } from './login'
 
 const makeEmailValidator = (): EmailValidator => {
@@ -114,5 +114,15 @@ describe('Login Controller', () => {
     const httpResponse = await sut.handle(makeFakeRequest())
 
     expect(httpResponse).toEqual(unauthorized())
+  })
+
+  test('Should return 500 if Authentication throws', async () => {
+    const { sut, emailValidatorStub } = makeSut()
+    jest.spyOn(emailValidatorStub, 'isValid').mockImplementationOnce(() => {
+      throw new Error()
+    })
+    const httpResponse = await sut.handle(makeFakeRequest())
+
+    expect(httpResponse).toEqual(serverError(new Error()))
   })
 })
