+++ src/presentation/controllers/login/login.ts
@@ -1,7 +1,6 @@
-import { Authentication } from '../../../domain/useCases/authentication'
-import { InvalidParamError, MissingParamError } from '../../errors'
+import { Controller, EmailValidator, HttpRequest, HttpResponse, Authentication } from './login-protocols'
 import { badRequest, serverError, unauthorized } from '../../helpers/http-helper'
-import { Controller, EmailValidator, HttpRequest, HttpResponse } from '../../protocols'
+import { InvalidParamError, MissingParamError } from '../../errors'
 
 export class LoginController implements Controller {
   private readonly emailValidator: EmailValidator
